+++ ThomasSite01/Program.cs
@@ -34,27 +34,64 @@ namespace ThomasSite01
                 }
                 else
                 {
-                    var foundCoffeeShops = coffeeShops
-                    .Where(x => x.Location.StartsWith(line, StringComparison.OrdinalIgnoreCase))
-                    .ToList();
-                    if (foundCoffeeShops.Count == 0)
+                    var commandHandler = new CoffeeShopCommandHandler(coffeeShops, line);  //ctrl. creates a class in a new file and we need to instantiate from this new class just created : CoffeeShopCommandHandler
+                    commandHandler.HandleCommand();                                        //ctrl. create a new method in the CoffeeShopCommandHandler newly created
+                    ///Here is the new class CoffeeShopCommandHandler
+                    /*
+                    using DataAccessSite07.Model;
+                    using System.Collections.Generic;
+
+                    namespace ThomasSite01
                     {
-                        Console.WriteLine($"> Command '{line}' not found");
-                    }
-                    else if (foundCoffeeShops.Count == 1)
-                    {
-                        var coffeeShop = foundCoffeeShops.Single();
-                        Console.WriteLine($"> Location: {coffeeShop.Location}");
-                        Console.WriteLine($"> Beans in stock: {coffeeShop.BeansInStockInKg} Kg");
-                    }
-                    else
-                    {
-                        Console.WriteLine($"> Multiple matching coffee shop command found:");
-                        foreach (var coffeeType in foundCoffeeShops)
+                        internal class CoffeeShopCommandHandler
                         {
-                            Console.WriteLine($"> {coffeeType.Location}");
+                            private IEnumerable<CoffeeShop> coffeeShops;
+                            private string line;
+
+                            public CoffeeShopCommandHandler(IEnumerable<CoffeeShop> coffeeShops, string line)
+                            {
+                                this.coffeeShops = coffeeShops;
+                                this.line = line;
+                            }
                         }
                     }
+                    ///added to the new method HandleCommand() to generate via ctrl. in the new class CoffeeShopCommandHandler
+                    using DataAccessSite07.Model;
+                    +++
+                    using System;
+                    ...
+                                this.line = line;
+                            }
+                    +++
+                            internal void HandleCommand()
+                            {
+                                throw new NotImplementedException();
+                            }
+                    */
+
+
+                    ///Need to MOVE because it got to big: using two inputs coffeeShops and line from this program.cs
+                    //var foundCoffeeShops = coffeeShops
+                    //.Where(x => x.Location.StartsWith(line, StringComparison.OrdinalIgnoreCase))
+                    //.ToList();
+                    //if (foundCoffeeShops.Count == 0)
+                    //{
+                    //    Console.WriteLine($"> Command '{line}' not found");
+                    //}
+                    //else if (foundCoffeeShops.Count == 1)
+                    //{
+                    //    var coffeeShop = foundCoffeeShops.Single();
+                    //    Console.WriteLine($"> Location: {coffeeShop.Location}");
+                    //    Console.WriteLine($"> Beans in stock: {coffeeShop.BeansInStockInKg} Kg");
+                    //}
+                    //else
+                    //{
+                    //    Console.WriteLine($"> Multiple matching coffee shop command found:");
+                    //    foreach (var coffeeType in foundCoffeeShops)
+                    //    {
+                    //        Console.WriteLine($"> {coffeeType.Location}");
+                    //    }
+                    //}
 
                 }
